getDEVUID/requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
openpyxl>=3.0.0
pexpect>=4.8.0 